Move event() and property() back into the public API because extensibility and both options accept a function in the definition.
