+++ test/unit.js
@@ -1,8 +1,10 @@
 import './boot';
 import './unit/api/create';
 import './unit/api/emit';
+import './unit/api/event';
 import './unit/api/init';
 import './unit/api/no-conflict';
+import './unit/api/property';
 import './unit/attributes';
 import './unit/constructor';
 import './unit/document-observer';
@@ -10,8 +12,6 @@ import './unit/dom';
 import './unit/extending';
 import './unit/ignoring';
 import './unit/lifecycle';
-import './unit/lifecycle/events';
-import './unit/lifecycle/property';
 import './unit/registration';
 import './unit/registry';
 import './unit/templating';
